agentic_project/requirements.txt
PyYAML
fastapi
uvicorn[standard]
pydantic
langchain-core
langchain-google-vertexai
google-cloud-aiplatform
langchain
chromadb